requirements.txt
boto3==1.34.36
botocore==1.34.36
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
einops==0.6.1
filelock==3.13.1
fonttools==4.42.0
fsspec==2024.2.0
ftfy==6.1.3
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.0.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.23.5
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
tokenizers==0.15.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
transformers==4.37.2
typing_extensions==4.7.1
urllib3==1.26.18
wcwidth==0.2.13
zipp==3.16.2